Fix Aurora authentication and annotation backend

+++ backend/src/services/annotation.service.ts
@@ -1,4 +1,5 @@
 import annotationRepository from "../repositories/annotation.repository";
+import Task from "../models/task.model";
 
 class AnnotationService {
   async getAnnotations(taskId: string) {
@@ -10,7 +11,13 @@ class AnnotationService {
   }
 
   async createAnnotation(data: any) {
-    return annotationRepository.create(data);
+    const annotation = await annotationRepository.create(data);
+
+    await Task.findByIdAndUpdate(annotation.task, {
+      $inc: { annotationCount: 1 },
+    });
+
+    return annotation;
   }
 
   async updateAnnotation(id: string, data: any) {
@@ -18,7 +25,21 @@ class AnnotationService {
   }
 
   async deleteAnnotation(id: string) {
-    return annotationRepository.delete(id);
+    const annotation = await annotationRepository.findById(id);
+
+    if (!annotation) {
+      return null;
+    }
+
+    const deleted = await annotationRepository.delete(id);
+
+    if (deleted) {
+      await Task.findByIdAndUpdate(annotation.task, {
+        $inc: { annotationCount: -1 },
+      });
+    }
+
+    return deleted;
   }
 
   async getAnnotationCount(taskId: string) {
